Split GameView into StatusView and BoardView
This splits the View into two separate classes, which should
make it easier to create game options in the status view
without adding too many responsibilities to the GameView.

Extract StatusView from GameView

Rename GameView to BoardView
