+++ app/main.coffee
@@ -1,14 +1,16 @@
 $ = require('jquery')
 Backbone = require('backbone')
 Game = require('./models/Game')
-GameView = require('./views/GameView')
+BoardView = require('./views/BoardView')
+StatusView = require('./views/StatusView')
 Client = require('./models/Client')
 Router = Backbone.Router.extend(
   routes: '': 'main'
   main: ->
     client = new Client
     game = new Game(client: client)
-    gameView = new GameView(model: game)
+    boardView = new BoardView(model: game)
+    statusView = new StatusView(model: game)
 )
 $(document).ready ->
   router = new Router
