+++ app/views/BoardView.coffee
@@ -1,8 +1,8 @@
 Backbone = require('backbone')
 $ = require('jquery')
 
-GameView = Backbone.View.extend(
-  el: '#game'
+BoardView = Backbone.View.extend(
+  el: '#board'
   events:
     'click .spot': 'move'
     'click #play-again': 'resetGame'
@@ -20,28 +20,10 @@ GameView = Backbone.View.extend(
       @model.makeMove spotClicked.attr('id')
 
   render: ->
-    @displayGameStatus()
     for i in [0...9]
       marker = @model.get('board')[i]
       $('#' + i).html @getMarkerHTML(marker)
 
-  displayGameStatus: ->
-    text = @getStatusText(@model.get('status'))
-    $("#status").html(text)
-  
-  getStatusText: (status) ->
-    switch status
-      when "tie" then "It's a tie!"
-      when "player1Wins" then "X Wins!"
-      when "player2Wins" then "O Wins!"
-      when "in progress" then @getInProgressText()
-
-  getInProgressText: ->
-    if @model.get('isXTurn')
-      "Your turn!"
-    else 
-      "Computer is thinking..."
-      
   getMarkerHTML: (marker) ->
     htmlclass = if marker == 'X' then 'human-move' else 'computer-move'
     "<span class=#{htmlclass}>#{marker}</span>"
@@ -72,4 +54,4 @@ GameView = Backbone.View.extend(
         $spot.addClass 'enabled'
 )
 
-module.exports = GameView
+module.exports = BoardView
