+++ spec/fixtures/status.html
@@ -1,0 +1,5 @@
+  <h1>Welcome to Tic Tac Toe!</h1>
+  <section id="game">
+    <h3 id="status">Your turn!</h3>
+  </section>
+
